NHANES blood pressure data

The US National Health and Nutrition Examination Survey (NHANES) was designed to assess the health and nutritional status of the non-institutionalized US population and was conducted by the National Center for Health Statistics of the Centers for Disease Control and Prevention (1). Since 1999-2000, NHANES has been conducted in two-year cycles using a multistage probability sampling design to select participants. Each cycle is independent with different participants recruited.

nhanes

Format

A data frame with columns:

exam

NHANES exam: 2013-2014, 2015-2016, or 2017-2018

seqn

survey participant identifier

psu

primary sampling unit

strata

survey strata

wts_mec_2yr

2 year mobile examination weights

exam_status

exam status. Participants either completed both the NHANES interview and exam or just the interview.

age

participant's age, in years

sex

participant's sex

race_ethnicity

participant's race and ethnicity

education

participant's education

pregnant

pregancy status

bp_sys_mmhg

participant's systolic blood pressure, mm Hg

bp_dia_mmhg

participant's diastolic blood pressure, mm Hg

n_msr_sbp

the number of valid systolic blood pressure readings

n_msr_dbp

the number of valid diastolic blood pressure readings

bp_high_aware

was participant ever told they had high blood pressure by a medical professional?

meds_bp

is participant currently using medication to lower their blood pressure?

Source

NHANES website, https://www.cdc.gov/nchs/nhanes/index.htm

Note

Blood pressure measurements

The same protocol was followed to measure systolic and diastolic blood pressure (SBP and DBP) in each NHANES cycle. After survey participants had rested 5 minutes, their BP was measured by a trained physician using a mercury sphygmomanometer and an appropriately sized cuff. Three BP measurements were obtained at 30 second intervals.

References

National health and nutrition examination survey homepage, available at https://www.cdc.gov/nchs/nhanes/index.htm. Accessed on 09/07/2020.

Examples

nhanes

#> # A tibble: 29,400 x 17
#>    exam   seqn   psu strata wts_mec_2yr exam_status   age sex   race_ethnicity
#>    <chr> <dbl> <dbl>  <dbl>       <dbl> <chr>       <dbl> <chr> <chr>         
#>  1 exam~ 73557     1    112      13481. interview ~    69 male  black         
#>  2 exam~ 73558     1    108      24472. interview ~    54 male  white         
#>  3 exam~ 73559     1    109      57193. interview ~    72 male  white         
#>  4 exam~ 73560     2    109      55767. interview ~     9 male  white         
#>  5 exam~ 73561     2    116      65542. interview ~    73 fema~ white         
#>  6 exam~ 73562     1    111      25345. interview ~    56 male  hispanic      
#>  7 exam~ 73563     1    105      17914. interview ~     0 male  white         
#>  8 exam~ 73564     1    114      61759. interview ~    61 fema~ white         
#>  9 exam~ 73565     2    106          0  interview ~    42 male  other         
#> 10 exam~ 73566     1    112      17480. interview ~    56 fema~ white         
#> # ... with 29,390 more rows, and 8 more variables: education <fct>,
#> #   pregnant <chr>, bp_sys_mmhg <dbl>, bp_dia_mmhg <dbl>, n_msr_sbp <int>,
#> #   n_msr_dbp <int>, bp_high_aware <chr>, meds_bp <chr>