Expressive rounding for table values

Expressive rounding for table values

table_glue(..., rspec = NULL, .sep = "", .envir = parent.frame())

Arguments

...
strings to round and format. Multiple inputs are concatenated together. Named arguments are not supported.

rspec
a rounding_specification object. If no rspec is given, a default setting will round values to decimal places based on the magnitude of the values.

.sep
Separator used to separate elements

.envir
environment to evaluate each expression in.

Value

a character vector of length equal to the vectors supplied in ...

See also

Examples

x <- runif(10)
y <- runif(10)

table_glue("{x} / {y} = {x/y}")

#>  [1] "0.87460--06608 / 0.28989--22954 = 3.01698--48413"
#>  [2] "0.17494--06268 / 0.67838--04274 = 0.25787--98264"
#>  [3] "0.03424--13327 / 0.73531--95988 = 0.04656--65987"
#>  [4] "0.32038--57308 / 0.19595--67331 = 1.63498--19973"
#>  [5] "0.40232--82384 / 0.98053--96746 = 0.41031--30641"
#>  [6] "0.19566--98347 / 0.74152--15292 = 0.26387--61345"
#>  [7] "0.40353--81174 / 0.05144--62763 = 7.84387--41698"
#>  [8] "0.06366--14573 / 0.53021--24636 = 0.12006--78250"
#>  [9] "0.38870--13132 / 0.69582--38790 = 0.55862--02557"
#> [10] "0.97554--78352 / 0.68855--60034 = 1.41680--24539"

table_glue("{x}", "({100 * y}%)", .sep = ' ')

#>  [1] "0.87460--06608 (28.98922--95375%)" "0.17494--06268 (67.83804--27416%)"
#>  [3] "0.03424--13327 (73.53195--98811%)" "0.32038--57308 (19.59567--33078%)"
#>  [5] "0.40232--82384 (98.05396--74638%)" "0.19566--98347 (74.15215--29155%)"
#>  [7] "0.40353--81174 (5.14462--76251%)"  "0.06366--14573 (53.02124--63578%)"
#>  [9] "0.38870--13132 (69.58238--78981%)" "0.97554--78352 (68.85560--03384%)"

df = data.frame(x = 1:10, y=1:10)

table_glue("{x} / {y} = {as.integer(x/y)}", .envir = df)

#>  [1] "1 / 1 = 1"   "2 / 2 = 1"   "3 / 3 = 1"   "4 / 4 = 1"   "5 / 5 = 1"  
#>  [6] "6 / 6 = 1"   "7 / 7 = 1"   "8 / 8 = 1"   "9 / 9 = 1"   "10 / 10 = 1"

table_glue("{x} / {y} = {as.integer(x/y)}", .envir = NULL)

#> Error in ..f(x): object 'x' not found

with(df, table_glue("{x} / {y} = {as.integer(x/y)}"))

#>  [1] "1 / 1 = 1"   "2 / 2 = 1"   "3 / 3 = 1"   "4 / 4 = 1"   "5 / 5 = 1"  
#>  [6] "6 / 6 = 1"   "7 / 7 = 1"   "8 / 8 = 1"   "9 / 9 = 1"   "10 / 10 = 1"

mtcars$car <- rownames(mtcars)
# use the default rounding specification
table_glue(
  "the {car} gets ~{mpg} miles/gallon and weighs ~{wt} thousand lbs",
  .envir = mtcars[1:3, ]
)

#> [1] "the Mazda RX4 gets ~21.00000--00000 miles/gallon and weighs ~2.62000--00000 thousand lbs"    
#> [2] "the Mazda RX4 Wag gets ~21.00000--00000 miles/gallon and weighs ~2.87500--00000 thousand lbs"
#> [3] "the Datsun 710 gets ~22.80000--00000 miles/gallon and weighs ~2.32000--00000 thousand lbs"   

# use your own rounding specification
rspec <- round_spec()
rspec <- round_using_decimal(rspec, digits = 1)

table_glue(
  "the {car} gets ~{mpg} miles/gallon and weighs ~{wt} thousand lbs",
  rspec = rspec,
  .envir = mtcars[1:3, ]
)

#> [1] "the Mazda RX4 gets ~21.0 miles/gallon and weighs ~2.6 thousand lbs"    
#> [2] "the Mazda RX4 Wag gets ~21.0 miles/gallon and weighs ~2.9 thousand lbs"
#> [3] "the Datsun 710 gets ~22.8 miles/gallon and weighs ~2.3 thousand lbs"