Expressive rounding for table values

Expressive rounding for table values

table_glue(..., rspec = NULL, .sep = "", .envir = parent.frame())

Arguments

...
strings to round and format. Multiple inputs are concatenated together. Named arguments are not supported.

rspec
a rounding_specification object. If no rspec is given, a default setting will round values to decimal places based on the magnitude of the values.

.sep
Separator used to separate elements

.envir
environment to evaluate each expression in.

Value

a character vector of length equal to the vectors supplied in ...

See also

Examples

x <- runif(10)
y <- runif(10)

table_glue("{x} / {y} = {x/y}")

#>  [1] "0.87 / 0.29 = 3.0"  "0.17 / 0.68 = 0.26" "0.03 / 0.74 = 0.05"
#>  [4] "0.32 / 0.20 = 1.6"  "0.40 / 0.98 = 0.41" "0.20 / 0.74 = 0.26"
#>  [7] "0.40 / 0.05 = 7.8"  "0.06 / 0.53 = 0.12" "0.39 / 0.70 = 0.56"
#> [10] "0.98 / 0.69 = 1.4" 

table_glue("{x}", "({100 * y}%)", .sep = ' ')

#>  [1] "0.87 (29%)"  "0.17 (68%)"  "0.03 (74%)"  "0.32 (20%)"  "0.40 (98%)" 
#>  [6] "0.20 (74%)"  "0.40 (5.1%)" "0.06 (53%)"  "0.39 (70%)"  "0.98 (69%)" 

df = data.frame(x = 1:10, y=1:10)

table_glue("{x} / {y} = {as.integer(x/y)}", .envir = df)

#>  [1] "1 / 1 = 1"   "2 / 2 = 1"   "3 / 3 = 1"   "4 / 4 = 1"   "5 / 5 = 1"  
#>  [6] "6 / 6 = 1"   "7 / 7 = 1"   "8 / 8 = 1"   "9 / 9 = 1"   "10 / 10 = 1"

table_glue("{x} / {y} = {as.integer(x/y)}")

#>  [1] "0.87 / 0.29 = 3" "0.17 / 0.68 = 0" "0.03 / 0.74 = 0" "0.32 / 0.20 = 1"
#>  [5] "0.40 / 0.98 = 0" "0.20 / 0.74 = 0" "0.40 / 0.05 = 7" "0.06 / 0.53 = 0"
#>  [9] "0.39 / 0.70 = 0" "0.98 / 0.69 = 1"

with(df, table_glue("{x} / {y} = {as.integer(x/y)}"))

#>  [1] "1 / 1 = 1"   "2 / 2 = 1"   "3 / 3 = 1"   "4 / 4 = 1"   "5 / 5 = 1"  
#>  [6] "6 / 6 = 1"   "7 / 7 = 1"   "8 / 8 = 1"   "9 / 9 = 1"   "10 / 10 = 1"

mtcars$car <- rownames(mtcars)
# use the default rounding specification
table_glue(
  "the {car} gets ~{mpg} miles/gallon and weighs ~{wt} thousand lbs",
  .envir = mtcars[1:3, ]
)

#> [1] "the Mazda RX4 gets ~21 miles/gallon and weighs ~2.6 thousand lbs"    
#> [2] "the Mazda RX4 Wag gets ~21 miles/gallon and weighs ~2.9 thousand lbs"
#> [3] "the Datsun 710 gets ~23 miles/gallon and weighs ~2.3 thousand lbs"   

# use your own rounding specification
rspec <- round_spec()
rspec <- round_using_decimal(rspec, digits = 1)

table_glue(
  "the {car} gets ~{mpg} miles/gallon and weighs ~{wt} thousand lbs",
  rspec = rspec,
  .envir = mtcars[1:3, ]
)

#> [1] "the Mazda RX4 gets ~21.0 miles/gallon and weighs ~2.6 thousand lbs"    
#> [2] "the Mazda RX4 Wag gets ~21.0 miles/gallon and weighs ~2.9 thousand lbs"
#> [3] "the Datsun 710 gets ~22.8 miles/gallon and weighs ~2.3 thousand lbs"